aquila/pyproject.toml
[project]
name = "aquila"
version = "0.1.0"
description = ""
authors = [
    {name = "almagribis",email = "almagribisadli@gmail.com"}
]
readme = "README.md"
packages = [{include = "aquila"}]
requires-python = ">=3.10,<4.0"
dependencies = [
    "langchain (==1.0.5)",
    "fastapi (>=0.119.0,<0.120.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langchain-community (>=0.4.1)",
    "streamlit (==1.51.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "langchain-google-genai (>=3.0.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.setuptools.packages.find]
where = ["."]
include = ["orion"]
namespaces = false